Word2Vec progress
Moved covariance from Matrix to Stats
