+++ scala212/src/test/scala/scalakittens/experiments/word2vec/SkipGramModelTest.scala
@@ -3,7 +3,8 @@ package scalakittens.experiments.word2vec
 import org.specs2.mutable.Specification
 
 import language.postfixOps
-import scalakittens.la.{Matrix, PCA}
+import scalakittens.la.{Basis, Matrix, PCA}
+import scalakittens.stats.AccumulatingMoments
 import scalakittens.{Good, IO, Strings}
 
 class SkipGramModelTest extends Specification {
@@ -19,7 +20,7 @@ class SkipGramModelTest extends Specification {
       source map scanner.scan match {
         case Good(st) =>
  
-          val model = SkipGramModel(st, 10, 0.25, 3, 10, 123456789L)
+          val model = SkipGramModel(st, 10, 0.25, 3, 100, 123456789L)
           
           model.run()
           val vectors = st.dictionary zip model.in
@@ -28,13 +29,15 @@ class SkipGramModelTest extends Specification {
           println("Frequent words")
           println((vectors takeRight 10).reverse mkString "\n")
           val size: Int = vectors.head._2.length
-          val (avg, cov) = Matrix.covariance(size, vectors map (_._2))
+          val acc = AccumulatingMoments(size).update(vectors map (_._2))
+          val avg = acc.avg
+          val cov = acc.covariance
           println(s"avg=$avg")
           println(s"covariance=\n$cov\n\n")
           val Some(eigens) = PCA.Iterations(0.001, 100).buildEigenVectors(cov, 10)
           println("\nEIGENVALUES:\n")
           println(eigens map (_._1))
-          
+          val newBasis = Basis(avg, Matrix.ofColumns(avg.length, eigens.map (_._2).toArray))          
           ok
           
         case bad =>
